Fixing the code duplicate error
